requirements.txt
Django==5.1.2
Pillow>=10
mysqlclient>=2.2    
weasyprint==66.0
django-environ>=0.11
